requirements.txt
torch==1.8.2
torchvision==0.9.2
flask==2.1.3
pillow==9.1.0
